--- a/examples/experimental/sun/src/app.js
+++ b/examples/experimental/sun/src/app.js
@@ -4,7 +4,6 @@ import {StaticMap} from 'react-map-gl';
 import DeckGL from '@deck.gl/react';
 import {AmbientLight, DirectionalLight, LightingEffect} from '@deck.gl/core';
 import {SolidPolygonLayer} from '@deck.gl/layers';
-import {PhongMaterial} from '@luma.gl/core';
 
 // Set your mapbox token here
 const MAPBOX_TOKEN = process.env.MapboxAccessToken; // eslint-disable-line
@@ -34,12 +33,12 @@ const dirLight1 = new DirectionalLight({
 
 const lightingEffect = new LightingEffect({ambientLight, dirLight0, dirLight1});
 
-const material = new PhongMaterial({
+const material = {
   ambient: 0.1,
   diffuse: 0.6,
   shininess: 32,
   specularColor: [60, 64, 70]
-});
+};
 
 const landCover = [[[-74.0, 40.7], [-74.02, 40.7], [-74.02, 40.72], [-74.0, 40.72]]];