--- a/modules/core/src/effects/lighting/camera-light.js
+++ b/modules/core/src/effects/lighting/camera-light.js
@@ -1,4 +1,4 @@
-import {PointLight} from './lights';
+import {PointLight} from './point-light';
 import {getUniformsFromViewport} from '../../shaderlib/project/viewport-uniforms';
 
 export default class CameraLight extends PointLight {