--- a/modules/core/src/index.js
+++ b/modules/core/src/index.js
@@ -30,7 +30,9 @@ export {COORDINATE_SYSTEM} from './lib/constants';
 
 // Effects
 export {default as LightingEffect} from './effects/lighting/lighting-effect';
-export {AmbientLight, PointLight, DirectionalLight} from './effects/lighting/lights';
+export {AmbientLight} from './effects/lighting/ambient-light';
+export {DirectionalLight} from './effects/lighting/directional-light';
+export {PointLight} from './effects/lighting/point-light';
 export {default as _CameraLight} from './effects/lighting/camera-light';
 export {default as _SunLight} from './effects/lighting/sun-light';
 export {default as PostProcessEffect} from './effects/post-process-effect';