--- a/modules/core/src/lib/deck.js
+++ b/modules/core/src/lib/deck.js
@@ -37,7 +37,7 @@ import {
   setParameters,
   lumaStats
 } from '@luma.gl/core';
-import {Timeline} from '@luma.gl/addons';
+import {Timeline} from '@luma.gl/engine';
 import {Stats} from 'probe.gl';
 import {EventManager} from 'mjolnir.js';