--- a/modules/core/src/lib/seer-integration.js
+++ b/modules/core/src/lib/seer-integration.js
@@ -127,14 +127,6 @@ function logPayload(layer) {
       const attrs = layer.getAttributeManager().getAttributes();
       data.push({path: 'objects.attributes', data: attrs});
     }
-    // TODO: Seer currently only handles single model layers
-    if (layer.state.model) {
-      layer.state.model.setProps({timerQueryEnabled: true});
-      const {lastFrameTime} = layer.state.model.stats;
-      if (lastFrameTime) {
-        badges.push(`${(lastFrameTime * 1000).toFixed(0)}μs`);
-      }
-    }
   }
 
   data.push({path: 'badges', data: badges});