--- a/modules/core/src/shaderlib/index.js
+++ b/modules/core/src/shaderlib/index.js
@@ -19,7 +19,7 @@
 // THE SOFTWARE.
 
 import {ProgramManager} from '@luma.gl/core';
-import {gouraudlighting as gouraudLighting, phonglighting as phongLighting} from '@luma.gl/core';
+import {gouraudLighting, phongLighting} from '@luma.gl/core';
 import geometry from './misc/geometry';
 import project from './project/project';
 import project32 from './project32/project32';