--- a/modules/mesh-layers/src/scenegraph-layer/scenegraph-layer.js
+++ b/modules/mesh-layers/src/scenegraph-layer/scenegraph-layer.js
@@ -20,7 +20,7 @@
 
 import {Layer, project32, picking} from '@deck.gl/core';
 import {ScenegraphNode, isWebGL2, pbr, log} from '@luma.gl/core';
-import {createGLTFObjects} from '@luma.gl/addons';
+import {ScenegraphNode, createGLTFObjects} from '@luma.gl/addons';
 import GL from '@luma.gl/constants';
 import {waitForGLTFAssets} from './gltf-utils';