--- a/modules/mesh-layers/src/simple-mesh-layer/simple-mesh-layer.js
+++ b/modules/mesh-layers/src/simple-mesh-layer/simple-mesh-layer.js
@@ -81,12 +81,6 @@ function getGeometry(data) {
 }
 
 const DEFAULT_COLOR = [0, 0, 0, 255];
-const defaultMaterial = {
-  ambient: 0.35,
-  diffuse: 0.6,
-  shininess: 32,
-  specularColor: [30, 30, 30]
-};
 
 const defaultProps = {
   mesh: {value: null, type: 'object', async: true},
@@ -104,7 +98,7 @@ const defaultProps = {
   // of the original mesh.
   wireframe: false,
   // Optional material for 'lighting' shader module
-  material: defaultMaterial,
+  material: true,
   getPosition: {type: 'accessor', value: x => x.position},
   getColor: {type: 'accessor', value: DEFAULT_COLOR},