Main changes: 
- Update imports
- Materials defined as data objects
  - Material classes were removed in luma 8
- Define light classes fully. 
  - The deck classes subclassed luma classes which were removed. Full definition is now in deck.gl
- Rename `gouraudlighting`/`phonglighting` to `gouraudLighting`/`phongLighting` (fix case)

Perf improvement on the stress test is not as large as I was hoping (it was ~60% on the luma stress test), but I still haven't hit the state tracking which still shows up as the biggest bottleneck.

|         |Before|After|Improvement|
|---------|------|-----|-----------|
|CPU Time |75ms  |70ms |5ms (7%)   |
|GPU Time |17ms  |17ms |0ms (0%)   |